fix: update Makefile to use #SecurityInsights schema
Currently the `lintyml` step is a no-op, because it doesn't explicitly
validate against the #SecurityInsights schema.

Fix that by specifying that vet should validate non-CUE files agains
that schema. This now aligns the `cue vet` checks here with the
si-tooling usage of the #SecurityInsights Go generated code:

https://github.com/ossf/si-tooling/blob/88ed8ab93e3fe1f4aab9ccabb5452942e0325a53/v2/si/security_insights.go#L67

Also use `cue vet` to validate the schema itself. This is the more
canonical way to validate a schema.

Signed-off-by: Paul Jolly <paul@myitcv.io>